+++ minidom/src/element.rs
@@ -17,22 +17,60 @@ use crate::error::{Error, Result};
 use crate::namespaces::NSChoice;
 use crate::node::Node;
 use crate::prefixes::{Namespace, Prefix, Prefixes};
+use crate::tree_builder::TreeBuilder;
 
 use std::collections::{btree_map, BTreeMap};
-use std::io::Write;
+use std::convert::{TryFrom, TryInto};
+use std::io::{BufRead, Write};
+use std::sync::Arc;
 
 use std::borrow::Cow;
 use std::str;
 
-use quick_xml::events::{BytesDecl, BytesEnd, BytesStart, Event};
-use quick_xml::Reader as EventReader;
-use quick_xml::Writer as EventWriter;
-
-use std::io::BufRead;
+use rxml::writer::{Encoder, Item, TrackNamespace};
+use rxml::{EventRead, Lexer, PullDriver, RawParser, XmlVersion};
 
 use std::str::FromStr;
 
 use std::slice;
+
+fn encode_and_write<W: Write, T: rxml::writer::TrackNamespace>(
+    item: Item<'_>,
+    enc: &mut Encoder<T>,
+    mut w: W,
+) -> rxml::Result<()> {
+    let mut buf = rxml::bytes::BytesMut::new();
+    enc.encode_into_bytes(item, &mut buf)
+        .expect("encoder driven incorrectly");
+    w.write_all(&buf[..])?;
+    Ok(())
+}
+
+/// Wrapper around a [`std::io::Write`] and an [`rxml::writer::Encoder`], to
+/// provide a simple function to write an rxml Item to a writer.
+pub struct CustomItemWriter<W, T> {
+    writer: W,
+    encoder: Encoder<T>,
+}
+
+impl<W: Write> CustomItemWriter<W, rxml::writer::SimpleNamespaces> {
+    pub(crate) fn new(writer: W) -> Self {
+        Self {
+            writer,
+            encoder: Encoder::new(),
+        }
+    }
+}
+
+impl<W: Write, T: rxml::writer::TrackNamespace> CustomItemWriter<W, T> {
+    pub(crate) fn write(&mut self, item: Item<'_>) -> rxml::Result<()> {
+        encode_and_write(item, &mut self.encoder, &mut self.writer)
+    }
+}
+
+/// Type alias to simplify the use for the default namespace tracking
+/// implementation.
+pub type ItemWriter<W> = CustomItemWriter<W, rxml::writer::SimpleNamespaces>;
 
 /// helper function to escape a `&[u8]` and replace all
 /// xml special characters (<, >, &, ', ") with their corresponding
@@ -82,11 +120,7 @@ pub fn escape(raw: &[u8]) -> Cow<[u8]> {
 pub struct Element {
     name: String,
     namespace: String,
-    /// This is only used when deserializing. If you have to use a custom prefix use
-    /// `ElementBuilder::prefix`.
-    #[allow(dead_code)]
-    prefix: Option<Prefix>,
-    /// Prefixes in the element
+    /// Namespace declarations
     pub prefixes: Prefixes,
     attributes: BTreeMap<String, String>,
     children: Vec<Node>,
@@ -104,8 +138,7 @@ impl FromStr for Element {
     type Err = Error;
 
     fn from_str(s: &str) -> Result<Element> {
-        let mut reader = EventReader::from_str(s);
-        Element::from_reader(&mut reader)
+        Element::from_reader(s.as_bytes())
     }
 }
 
@@ -122,28 +155,17 @@ impl PartialEq for Element {
     }
 }
 
-fn ensure_no_prefix<S: AsRef<str>>(s: &S) -> Result<()> {
-    match s.as_ref().split(':').count() {
-        1 => Ok(()),
-        _ => Err(Error::InvalidElement),
-    }
-}
-
 impl Element {
-    fn new<P: Into<Prefixes>>(
+    pub(crate) fn new<P: Into<Prefixes>>(
         name: String,
         namespace: String,
-        prefix: Option<Prefix>,
         prefixes: P,
         attributes: BTreeMap<String, String>,
         children: Vec<Node>,
     ) -> Element {
-        ensure_no_prefix(&name).unwrap();
-        // TODO: Return Result<Element> instead.
         Element {
             name,
             namespace,
-            prefix,
             prefixes: prefixes.into(),
             attributes,
             children,
@@ -174,7 +196,6 @@ impl Element {
                 name.as_ref().to_string(),
                 namespace.into(),
                 None,
-                None,
                 BTreeMap::new(),
                 Vec::new(),
             ),
@@ -199,7 +220,6 @@ impl Element {
         Element::new(
             name.into(),
             namespace.into(),
-            None,
             None,
             BTreeMap::new(),
             Vec::new(),
@@ -312,239 +332,101 @@ impl Element {
         namespace.into().compare(self.namespace.as_ref())
     }
 
-    /// Parse a document from an `EventReader`.
-    pub fn from_reader<R: BufRead>(reader: &mut EventReader<R>) -> Result<Element> {
-        let mut buf = Vec::new();
+    /// Parse a document from a `BufRead`.
+    pub fn from_reader<R: BufRead>(reader: R) -> Result<Element> {
+        let mut tree_builder = TreeBuilder::new();
+        let mut driver = PullDriver::wrap(reader, Lexer::new(), RawParser::new());
+        while let Some(event) = driver.read()? {
+            tree_builder.process_event(event)?;
 
-        let mut prefixes = BTreeMap::new();
-        let root: Element = loop {
-            let e = reader.read_event(&mut buf)?;
-            match e {
-                Event::Empty(ref e) | Event::Start(ref e) => {
-                    break build_element(reader, e, &mut prefixes)?;
-                }
-                Event::Eof => {
-                    return Err(Error::EndOfDocument);
-                }
-                Event::Comment { .. } => {
-                    return Err(Error::NoComments);
-                }
-                Event::Text { .. }
-                | Event::End { .. }
-                | Event::CData { .. }
-                | Event::Decl { .. }
-                | Event::PI { .. }
-                | Event::DocType { .. } => (), // TODO: may need more errors
-            }
-        };
-
-        let mut stack = vec![root];
-        let mut prefix_stack = vec![prefixes];
-
-        loop {
-            match reader.read_event(&mut buf)? {
-                Event::Empty(ref e) => {
-                    let mut prefixes = prefix_stack.last().unwrap().clone();
-                    let elem = build_element(reader, e, &mut prefixes)?;
-                    // Since there is no Event::End after, directly append it to the current node
-                    stack.last_mut().unwrap().append_child(elem);
-                }
-                Event::Start(ref e) => {
-                    let mut prefixes = prefix_stack.last().unwrap().clone();
-                    let elem = build_element(reader, e, &mut prefixes)?;
-                    stack.push(elem);
-                    prefix_stack.push(prefixes);
-                }
-                Event::End(ref e) => {
-                    if stack.len() <= 1 {
-                        break;
-                    }
-                    let prefixes = prefix_stack.pop().unwrap();
-                    let elem = stack.pop().unwrap();
-                    if let Some(to) = stack.last_mut() {
-                        // TODO: check whether this is correct, we are comparing &[u8]s, not &strs
-                        let elem_name = e.name();
-                        let mut split_iter = elem_name.splitn(2, |u| *u == 0x3A);
-                        let possible_prefix = split_iter.next().unwrap(); // Can't be empty.
-                        let opening_prefix = {
-                            let mut tmp: Option<Option<String>> = None;
-                            for (prefix, ns) in prefixes {
-                                if ns == elem.namespace {
-                                    tmp = Some(prefix.clone());
-                                    break;
-                                }
-                            }
-                            match tmp {
-                                Some(prefix) => prefix,
-                                None => return Err(Error::InvalidPrefix),
-                            }
-                        };
-                        match split_iter.next() {
-                            // There is a prefix on the closing tag
-                            Some(name) => {
-                                // Does the closing prefix match the opening prefix?
-                                match opening_prefix {
-                                    Some(prefix) if possible_prefix == prefix.as_bytes() => (),
-                                    _ => return Err(Error::InvalidElementClosed),
-                                }
-                                // Does the closing tag name match the opening tag name?
-                                if name != elem.name().as_bytes() {
-                                    return Err(Error::InvalidElementClosed);
-                                }
-                            }
-                            // There was no prefix on the closing tag
-                            None => {
-                                // Is there a prefix on the opening tag?
-                                if opening_prefix.is_some() {
-                                    return Err(Error::InvalidElementClosed);
-                                }
-                                // Does the opening tag name match the closing one?
-                                if possible_prefix != elem.name().as_bytes() {
-                                    return Err(Error::InvalidElementClosed);
-                                }
-                            }
-                        }
-                        to.append_child(elem);
-                    }
-                }
-                Event::Text(s) => {
-                    let text = s.unescape_and_decode(reader)?;
-                    if !text.is_empty() {
-                        let current_elem = stack.last_mut().unwrap();
-                        current_elem.append_text_node(text);
-                    }
-                }
-                Event::CData(s) => {
-                    let text = s.unescape_and_decode(&reader)?;
-                    if !text.is_empty() {
-                        let current_elem = stack.last_mut().unwrap();
-                        current_elem.append_text_node(text);
-                    }
-                }
-                Event::Eof => {
-                    break;
-                }
-                Event::Comment(_) => return Err(Error::NoComments),
-                Event::Decl { .. } | Event::PI { .. } | Event::DocType { .. } => (),
+            if let Some(root) = tree_builder.root.take() {
+                return Ok(root);
             }
         }
-        Ok(stack.pop().unwrap())
+        Err(Error::EndOfDocument)
+    }
+
+    /// Parse a document from a `BufRead`, allowing Prefixes to be specified. Useful to provide
+    /// knowledge of namespaces that would have been declared on parent elements not present in the
+    /// reader.
+    pub fn from_reader_with_prefixes<R: BufRead, P: Into<Prefixes>>(
+        reader: R,
+        prefixes: P,
+    ) -> Result<Element> {
+        let mut tree_builder = TreeBuilder::new().with_prefixes_stack(vec![prefixes.into()]);
+        let mut driver = PullDriver::wrap(reader, Lexer::new(), RawParser::new());
+        while let Some(event) = driver.read()? {
+            tree_builder.process_event(event)?;
+
+            if let Some(root) = tree_builder.root.take() {
+                return Ok(root);
+            }
+        }
+        Err(Error::EndOfDocument)
     }
 
     /// Output a document to a `Writer`.
     pub fn write_to<W: Write>(&self, writer: &mut W) -> Result<()> {
-        self.to_writer(&mut EventWriter::new(writer))
+        self.to_writer(&mut ItemWriter::new(writer))
     }
 
     /// Output a document to a `Writer`.
     pub fn write_to_decl<W: Write>(&self, writer: &mut W) -> Result<()> {
-        self.to_writer_decl(&mut EventWriter::new(writer))
+        self.to_writer_decl(&mut ItemWriter::new(writer))
     }
 
-    /// Output the document to quick-xml `Writer`
-    pub fn to_writer<W: Write>(&self, writer: &mut EventWriter<W>) -> Result<()> {
-        self.write_to_inner(writer, &mut BTreeMap::new())
+    /// Output the document to an `ItemWriter`
+    pub fn to_writer<W: Write>(&self, writer: &mut ItemWriter<W>) -> Result<()> {
+        self.write_to_inner(writer)
     }
 
-    /// Output the document to quick-xml `Writer`
-    pub fn to_writer_decl<W: Write>(&self, writer: &mut EventWriter<W>) -> Result<()> {
-        writer.write_event(Event::Decl(BytesDecl::new(b"1.0", Some(b"utf-8"), None)))?;
-        self.write_to_inner(writer, &mut BTreeMap::new())
+    /// Output the document to an `ItemWriter`
+    pub fn to_writer_decl<W: Write>(&self, writer: &mut ItemWriter<W>) -> Result<()> {
+        writer
+            .write(Item::XmlDeclaration(XmlVersion::V1_0))
+            .unwrap(); // TODO: error return
+        self.write_to_inner(writer)
     }
 
     /// Like `write_to()` but without the `<?xml?>` prelude
-    pub fn write_to_inner<W: Write>(
-        &self,
-        writer: &mut EventWriter<W>,
-        all_prefixes: &mut BTreeMap<Prefix, Namespace>,
-    ) -> Result<()> {
-        let local_prefixes: &BTreeMap<Option<String>, String> = self.prefixes.declared_prefixes();
-        // Element namespace
-        // If the element prefix hasn't been set yet via a custom prefix, add it.
-        let mut existing_self_prefix: Option<Option<String>> = None;
-        for (prefix, ns) in local_prefixes.iter().chain(all_prefixes.iter()) {
-            if ns == &self.namespace {
-                existing_self_prefix = Some(prefix.clone());
-            }
+    pub fn write_to_inner<W: Write>(&self, writer: &mut ItemWriter<W>) -> Result<()> {
+        for (prefix, namespace) in self.prefixes.declared_prefixes() {
+            assert!(writer.encoder.inner_mut().declare_fixed(
+                prefix.as_ref().map(|x| (&**x).try_into()).transpose()?,
+                Some(Arc::new(namespace.clone().try_into()?))
+            ));
         }
 
-        let mut all_keys = all_prefixes.keys().cloned();
-        let mut local_keys = local_prefixes.keys().cloned();
-        let self_prefix: (Option<String>, bool) = match existing_self_prefix {
-            // No prefix exists already for our namespace
-            None => {
-                if !local_keys.any(|p| p.is_none()) {
-                    // Use the None prefix if available
-                    (None, true)
-                } else {
-                    // Otherwise generate one. Check if it isn't already used, if so increase the
-                    // number until we find a suitable one.
-                    let mut prefix_n = 0u8;
-                    while all_keys.any(|p| p == Some(format!("ns{}", prefix_n))) {
-                        prefix_n += 1;
-                    }
-                    (Some(format!("ns{}", prefix_n)), true)
-                }
-            }
-            // Some prefix has already been declared (or is going to be) for our namespace. We
-            // don't need to declare a new one. We do however need to remember which one to use in
-            // the tag name.
-            Some(prefix) => (prefix, false),
+        let namespace = if self.namespace.len() == 0 {
+            None
+        } else {
+            Some(Arc::new(self.namespace.clone().try_into()?))
         };
+        writer.write(Item::ElementHeadStart(namespace, (*self.name).try_into()?))?;
 
-        let name = match self_prefix {
-            (Some(ref prefix), _) => Cow::Owned(format!("{}:{}", prefix, self.name)),
-            _ => Cow::Borrowed(&self.name),
-        };
-        let mut start = BytesStart::borrowed(name.as_bytes(), name.len());
+        for (key, value) in self.attributes.iter() {
+            let (prefix, name) = <&rxml::NameStr>::try_from(&**key)
+                .unwrap()
+                .split_name()
+                .unwrap();
+            let namespace = match prefix {
+                Some(prefix) => match writer.encoder.inner().lookup_prefix(Some(prefix)) {
+                    Ok(v) => Some(v),
+                    Err(rxml::writer::PrefixError::Undeclared) => return Err(Error::InvalidPrefix),
+                },
+                None => None,
+            };
+            writer.write(Item::Attribute(namespace, name, (&**value).try_into()?))?;
+        }
 
-        // Write self prefix if necessary
-        match self_prefix {
-            (Some(ref p), true) => {
-                let key = format!("xmlns:{}", p);
-                start.push_attribute((key.as_bytes(), self.namespace.as_bytes()));
-                all_prefixes.insert(self_prefix.0, self.namespace.clone());
-            }
-            (None, true) => {
-                let key = String::from("xmlns");
-                start.push_attribute((key.as_bytes(), self.namespace.as_bytes()));
-                all_prefixes.insert(self_prefix.0, self.namespace.clone());
-            }
-            _ => (),
-        };
-
-        // Custom prefixes/namespace sets
-        for (prefix, ns) in local_prefixes {
-            match all_prefixes.get(prefix) {
-                p @ Some(_) if p == prefix.as_ref() => (),
-                _ => {
-                    let key = match prefix {
-                        None => String::from("xmlns"),
-                        Some(p) => format!("xmlns:{}", p),
-                    };
-
-                    start.push_attribute((key.as_bytes(), ns.as_ref()));
-                    all_prefixes.insert(prefix.clone(), ns.clone());
-                }
+        if !self.children.is_empty() {
+            writer.write(Item::ElementHeadEnd)?;
+            for child in self.children.iter() {
+                child.write_to_inner(writer)?;
             }
         }
+        writer.write(Item::ElementFoot)?;
 
-        for (key, value) in &self.attributes {
-            start.push_attribute((key.as_bytes(), escape(value.as_bytes()).as_ref()));
-        }
-
-        if self.children.is_empty() {
-            writer.write_event(Event::Empty(start))?;
-            return Ok(());
-        }
-
-        writer.write_event(Event::Start(start))?;
-
-        for child in &self.children {
-            child.write_to_inner(writer, &mut all_prefixes.clone())?;
-        }
-
-        writer.write_event(Event::End(BytesEnd::borrowed(name.as_bytes())))?;
         Ok(())
     }
 
@@ -823,68 +705,18 @@ impl Element {
         })?;
         self.children.remove(idx).into_element()
     }
-}
 
-fn split_element_name<S: AsRef<str>>(s: S) -> Result<(Option<String>, String)> {
-    let name_parts = s.as_ref().split(':').collect::<Vec<&str>>();
-    match name_parts.len() {
-        2 => Ok((Some(name_parts[0].to_owned()), name_parts[1].to_owned())),
-        1 => Ok((None, name_parts[0].to_owned())),
-        _ => Err(Error::InvalidElement),
-    }
-}
-
-fn build_element<R: BufRead>(
-    reader: &EventReader<R>,
-    event: &BytesStart,
-    prefixes: &mut BTreeMap<Prefix, Namespace>,
-) -> Result<Element> {
-    let (prefix, name) = split_element_name(str::from_utf8(event.name())?)?;
-    let mut local_prefixes = BTreeMap::new();
-
-    let attributes = event
-        .attributes()
-        .map(|o| {
-            let o = o?;
-            let key = str::from_utf8(o.key)?.to_owned();
-            let value = o.unescape_and_decode_value(reader)?;
-            Ok((key, value))
-        })
-        .filter(|o| match *o {
-            Ok((ref key, ref value)) if key == "xmlns" => {
-                local_prefixes.insert(None, value.clone());
-                prefixes.insert(None, value.clone());
-                false
+    /// Remove the leading nodes up to the first child element and
+    /// return it
+    pub fn unshift_child(&mut self) -> Option<Element> {
+        while self.children.len() > 0 {
+            if let Some(el) = self.children.remove(0).into_element() {
+                return Some(el);
             }
-            Ok((ref key, ref value)) if key.starts_with("xmlns:") => {
-                local_prefixes.insert(Some(key[6..].to_owned()), value.to_owned());
-                prefixes.insert(Some(key[6..].to_owned()), value.to_owned());
-                false
-            }
-            _ => true,
-        })
-        .collect::<Result<BTreeMap<String, String>>>()?;
-
-    let namespace: &String = {
-        if let Some(namespace) = local_prefixes.get(&prefix) {
-            namespace
-        } else if let Some(namespace) = prefixes.get(&prefix) {
-            namespace
-        } else {
-            return Err(Error::MissingNamespace);
         }
-    };
 
-    Ok(Element::new(
-        name,
-        namespace.clone(),
-        // Note that this will always be Some(_) as we can't distinguish between the None case and
-        // Some(None). At least we make sure the prefix has a namespace associated.
-        Some(prefix),
-        local_prefixes,
-        attributes,
-        Vec::new(),
-    ))
+        None
+    }
 }
 
 /// An iterator over references to child elements of an `Element`.
@@ -1054,7 +886,6 @@ mod tests {
         let elem = Element::new(
             "name".to_owned(),
             "namespace".to_owned(),
-            None,
             (None, "namespace".to_owned()),
             BTreeMap::from_iter(vec![("name".to_string(), "value".to_string())].into_iter()),
             Vec::new(),
@@ -1068,9 +899,8 @@ mod tests {
 
     #[test]
     fn test_from_reader_simple() {
-        let xml = "<foo xmlns='ns1'></foo>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader);
+        let xml = b"<foo xmlns='ns1'></foo>";
+        let elem = Element::from_reader(&xml[..]);
 
         let elem2 = Element::builder("foo", "ns1").build();
 
@@ -1079,9 +909,8 @@ mod tests {
 
     #[test]
     fn test_from_reader_nested() {
-        let xml = "<foo xmlns='ns1'><bar xmlns='ns1' baz='qxx' /></foo>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader);
+        let xml = b"<foo xmlns='ns1'><bar xmlns='ns1' baz='qxx' /></foo>";
+        let elem = Element::from_reader(&xml[..]);
 
         let nested = Element::builder("bar", "ns1").attr("baz", "qxx").build();
         let elem2 = Element::builder("foo", "ns1").append(nested).build();
@@ -1091,9 +920,8 @@ mod tests {
 
     #[test]
     fn test_from_reader_with_prefix() {
-        let xml = "<foo xmlns='ns1'><prefix:bar xmlns:prefix='ns1' baz='qxx' /></foo>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader);
+        let xml = b"<foo xmlns='ns1'><prefix:bar xmlns:prefix='ns1' baz='qxx' /></foo>";
+        let elem = Element::from_reader(&xml[..]);
 
         let nested = Element::builder("bar", "ns1").attr("baz", "qxx").build();
         let elem2 = Element::builder("foo", "ns1").append(nested).build();
@@ -1103,9 +931,8 @@ mod tests {
 
     #[test]
     fn test_from_reader_split_prefix() {
-        let xml = "<foo:bar xmlns:foo='ns1'/>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader).unwrap();
+        let xml = b"<foo:bar xmlns:foo='ns1'/>";
+        let elem = Element::from_reader(&xml[..]).unwrap();
 
         assert_eq!(elem.name(), String::from("bar"));
         assert_eq!(elem.ns(), String::from("ns1"));
@@ -1119,41 +946,63 @@ mod tests {
     #[test]
     fn parses_spectest_xml() {
         // From: https://gitlab.com/lumi/minidom-rs/issues/8
-        let xml = r#"
-            <rng:grammar xmlns:rng="http://relaxng.org/ns/structure/1.0">
+        let xml = br#"<rng:grammar xmlns:rng="http://relaxng.org/ns/structure/1.0">
                 <rng:name xmlns:rng="http://relaxng.org/ns/structure/1.0"></rng:name>
             </rng:grammar>
         "#;
-        let mut reader = EventReader::from_str(xml);
-        let _ = Element::from_reader(&mut reader).unwrap();
+        let _ = Element::from_reader(&xml[..]).unwrap();
     }
 
     #[test]
     fn does_not_unescape_cdata() {
-        let xml = "<test xmlns='test'><![CDATA[&apos;&gt;blah<blah>]]></test>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader).unwrap();
+        let xml = b"<test xmlns='test'><![CDATA[&apos;&gt;blah<blah>]]></test>";
+        let elem = Element::from_reader(&xml[..]).unwrap();
         assert_eq!(elem.text(), "&apos;&gt;blah<blah>");
     }
 
     #[test]
     fn test_compare_all_ns() {
-        let xml = "<foo xmlns='foo' xmlns:bar='baz'><bar:meh xmlns:bar='baz' /></foo>";
-        let mut reader = EventReader::from_str(xml);
-        let elem = Element::from_reader(&mut reader).unwrap();
+        let xml = b"<foo xmlns='foo' xmlns:bar='baz'><bar:meh xmlns:bar='baz' /></foo>";
+        let elem = Element::from_reader(&xml[..]).unwrap();
 
         let elem2 = elem.clone();
 
-        let xml3 = "<foo xmlns='foo'><bar:meh xmlns:bar='baz'/></foo>";
-        let mut reader3 = EventReader::from_str(xml3);
-        let elem3 = Element::from_reader(&mut reader3).unwrap();
+        let xml3 = b"<foo xmlns='foo'><bar:meh xmlns:bar='baz'/></foo>";
+        let elem3 = Element::from_reader(&xml3[..]).unwrap();
 
-        let xml4 = "<prefix:foo xmlns:prefix='foo'><bar:meh xmlns:bar='baz'/></prefix:foo>";
-        let mut reader4 = EventReader::from_str(xml4);
-        let elem4 = Element::from_reader(&mut reader4).unwrap();
+        let xml4 = b"<prefix:foo xmlns:prefix='foo'><bar:meh xmlns:bar='baz'/></prefix:foo>";
+        let elem4 = Element::from_reader(&xml4[..]).unwrap();
 
         assert_eq!(elem, elem2);
         assert_eq!(elem, elem3);
         assert_eq!(elem, elem4);
+    }
+
+    #[test]
+    fn test_from_reader_with_prefixes() {
+        let xml = b"<foo><bar xmlns='baz'/></foo>";
+        let elem =
+            Element::from_reader_with_prefixes(&xml[..], String::from("jabber:client")).unwrap();
+
+        let xml2 = b"<foo xmlns='jabber:client'><bar xmlns='baz'/></foo>";
+        let elem2 = Element::from_reader(&xml2[..]).unwrap();
+
+        assert_eq!(elem, elem2);
+    }
+
+    #[test]
+    fn failure_with_duplicate_namespace() {
+        let _: Element = r###"<?xml version="1.0" encoding="UTF-8"?>
+            <wsdl:definitions
+                    xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/"
+                    xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+                <wsdl:types>
+                    <xsd:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
+                    </xsd:schema>
+                </wsdl:types>
+            </wsdl:definitions>
+        "###
+        .parse()
+        .unwrap();
     }
 }