+++ minidom/CHANGELOG.md
@@ -1,3 +1,18 @@
+Version 0.15.2, released 2023-05-13:
+  * Changes
+    * Fix a memory corruption on closing tags for elements with a name longer
+      than 24 bytes
+    * Only enable the mt features of rxml, we don’t need any additional one
+
+Version 0.15.1, released 2023-01-15:
+  * Changes
+    * Add `Element::from_reader_with_prefixes`
+    * (#44) Add test ensuring parsing two namespaces resolv
+
+Version 0.15.0, released 2022-07-13:
+  * Changes
+    * Drop quick-xml dependency (astro1, jssfr)
+
 Version 0.14.0, released 2022-03-07:
   * Changes
     * Bump quick-xml dependency (thanks eijebong!)